+++ src/Admin/AdminConsole/Models/ProviderPlanViewModel.cs
@@ -1,0 +1,26 @@
+﻿using Bit.Core.Billing.Entities;
+
+namespace Bit.Admin.AdminConsole.Models;
+
+public class ProviderPlanViewModel
+{
+    public string Name { get; set; }
+    public int PurchasedSeats { get; set; }
+    public int AssignedSeats { get; set; }
+    public int UsedSeats { get; set; }
+    public int RemainingSeats { get; set; }
+
+    public ProviderPlanViewModel(
+        string name,
+        ProviderPlan providerPlan,
+        int usedSeats)
+    {
+        var purchasedSeats = providerPlan.SeatMinimum ?? 0 + providerPlan.PurchasedSeats ?? 0;
+
+        Name = name;
+        PurchasedSeats = purchasedSeats;
+        AssignedSeats = providerPlan.AllocatedSeats ?? 0;
+        UsedSeats = usedSeats;
+        RemainingSeats = purchasedSeats - AssignedSeats;
+    }
+}
