+++ src/Admin/AdminConsole/Models/ProviderViewModel.cs
@@ -63,27 +63,4 @@ public class ProviderViewModel
     public IEnumerable<ProviderUserUserDetails> ProviderAdmins { get; set; }
     public IEnumerable<ProviderOrganizationOrganizationDetails> ProviderOrganizations { get; set; }
     public List<ProviderPlanViewModel> ProviderPlanViewModels { get; set; } = [];
-
-    public class ProviderPlanViewModel
-    {
-        public string Name { get; set; }
-        public int PurchasedSeats { get; set; }
-        public int AssignedSeats { get; set; }
-        public int UsedSeats { get; set; }
-        public int RemainingSeats { get; set; }
-
-        public ProviderPlanViewModel(
-            string name,
-            ProviderPlan providerPlan,
-            int usedSeats)
-        {
-            var purchasedSeats = providerPlan.SeatMinimum ?? 0 + providerPlan.PurchasedSeats ?? 0;
-
-            Name = name;
-            PurchasedSeats = purchasedSeats;
-            AssignedSeats = providerPlan.AllocatedSeats ?? 0;
-            UsedSeats = usedSeats;
-            RemainingSeats = purchasedSeats - AssignedSeats;
-        }
-    }
 }
