+++ src/Admin/AdminConsole/Views/Providers/ProviderPlans.cshtml
@@ -1,4 +1,4 @@
-@model List<ProviderViewModel.ProviderPlanViewModel>
+@model List<ProviderPlanViewModel>
 @foreach (var plan in Model)
 {
     <h2>@plan.Name</h2>
